UI/requirements.txt
numpy
matplotlib
pandas
pyserial
dash 
dash-bootstrap-components
plotly
